Isolate windows code
